requirements.txt
paramiko>=3.4.0
pyyaml>=6.0.1
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
psutil>=5.9.6
colorama>=0.4.6
tqdm>=4.66.1
watchdog>=3.0.0
structlog>=23.2.0
tenacity>=8.2.3
aiohttp>=3.8.0
dnspython>=2.4.0 
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
llama-index>=0.9.0
jedi>=0.19.1
astunparse>=1.6.3 